+++ reana_db/models.py
@@ -14,7 +14,9 @@ import enum
 import math
 import uuid
 from datetime import datetime
+import logging
 
+from reana_commons.config import REANA_RUNTIME_KUBERNETES_KEEP_ALIVE_JOBS_WITH_STATUSES
 from reana_commons.utils import get_disk_usage
 from sqlalchemy import (
     BigInteger,
@@ -44,6 +46,7 @@ from reana_db.utils import (
     get_default_quota_resource,
     update_users_disk_quota,
 )
+
 
 Base = declarative_base()
 
@@ -283,7 +286,28 @@ class UserToken(Base, Timestamp):
     type_ = Column(Enum(UserTokenType), nullable=False)
 
 
-class RunStatus(enum.Enum):
+class CleanUpDependingOnStatusMixin:
+    """Mixin to determine whether to clean up jobs for REANA status enums."""
+
+    @classmethod
+    def should_cleanup_job(cls, job_status):
+        """Determine if a job/workflow should be cleaned up depending on its status."""
+        job_status_name = (
+            job_status.name if isinstance(job_status, enum.Enum) else job_status
+        )
+        keep_on_status_set = set(REANA_RUNTIME_KUBERNETES_KEEP_ALIVE_JOBS_WITH_STATUSES)
+        all_statuses_list = [s.name for s in cls]
+        if not keep_on_status_set.issubset(all_statuses_list):
+            logging.warning(
+                "The configuration variable REANA_RUNTIME_KUBERNETES_KEEP_ALIVE_JOBS_WITH_STATUSES contains "
+                "unknown statuses {} which will be ignored, possibly causing jobs not to be cleaned up.".format(
+                    keep_on_status_set.difference(all_statuses_list)
+                )
+            )
+        return not (job_status_name in keep_on_status_set)
+
+
+class RunStatus(CleanUpDependingOnStatusMixin, enum.Enum):
     """Enumeration of possible run statuses."""
 
     created = 0
@@ -322,7 +346,7 @@ ALLOWED_WORKFLOW_STATUS_TRANSITIONS = [
 ]
 
 
-class JobStatus(enum.Enum):
+class JobStatus(CleanUpDependingOnStatusMixin, enum.Enum):
     """Enumeration of possible job statuses."""
 
     created = 0